Move isDefined to global utils
