--- src/ui/scripts/utils/isDefined.js
@@ -1,1 +1,0 @@
-export default (_) => _ != null
