+++ src/utils/isDefined.js
@@ -1,0 +1,3 @@
+'use strict'
+
+module.exports = (_) => _ != null
